+++ tests/setup.ps1
@@ -18,7 +18,8 @@ if ($Unit) {
  "id": "1234",
  "access_token": "token",
  "sessionToken": "token",
- "profile":{"name": "test"},
+ "profile":{"name": "test", "label":"test"},
+ "label": "OktaPosh-test-app",
  "_links": {
      "self": {
          "href" : "test/123"