+++ tests/unit/User.tests.ps1
@@ -38,7 +38,7 @@ Describe "User" {
 
         $saltedValue = $value + $saltValue
 
-        $pwValue = (New-Object 'System.Security.Cryptography.SHA256Managed').ComputeHash($saltedValue)
+        $pwValue = [System.Security.Cryptography.SHA256]::Create().ComputeHash($saltedValue)
 
         $passwordHash = @{
             hash = @{